requirements.txt
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dotenv==0.20.0
requests==2.28.1
slack-bolt==1.14.0
slack-sdk==3.17.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
